backend/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
pydantic
apscheduler==3.10.4
requests
google-auth
google-api-python-client
google-auth-httplib2
google-cloud-aiplatform
